Move main executable to a single file
